+++ __tests__/app/entry.js
@@ -3,6 +3,8 @@ import Application, { Text, Socket } from '../../src/app';
 
 import dictionaryCzech from './res/text_cs-CZ.json';
 
+import 'bootstrap/dist/css/bootstrap.css';
+
 TagManager.initialize({
     gtmId: 'GTM-KQ3XSLZ',
 });